+++ spring-cloud-tencent-examples/metadata-transfer-example/metadata-frontend/src/main/java/com/tencent/cloud/metadata/service/frontend/MetadataFrontendController.java
@@ -69,7 +69,7 @@ public class MetadataFrontendController {
 
 		// Get Custom Metadata From Context
 		MetadataContext context = MetadataContextHolder.get();
-		Map<String, String> customMetadataMap = context.getFragmentContext(MetadataContext.FRAGMENT_TRANSITIVE);
+		Map<String, String> customMetadataMap = context.getTransitiveMetadata();
 
 		customMetadataMap.forEach((key, value) -> {
 			LOG.info("Metadata Middle Custom Metadata (Key-Value): {} : {}", key, value);
@@ -116,7 +116,7 @@ public class MetadataFrontendController {
 
 		// Get Custom Metadata From Context
 		MetadataContext context = MetadataContextHolder.get();
-		Map<String, String> customMetadataMap = context.getFragmentContext(MetadataContext.FRAGMENT_TRANSITIVE);
+		Map<String, String> customMetadataMap = context.getTransitiveMetadata();
 
 		customMetadataMap.forEach((key, value) -> {
 			LOG.info("Metadata Middle Custom Metadata (Key-Value): {} : {}", key, value);